Feat/some ui (#27)
* ConfirmDialog

* Modify relabel page layout

+++ VisionOnEdge/EdgeSolution/modules/WebModule/ui/src/components/CameraDetails/CameraDetailInfo.tsx
@@ -1,7 +1,8 @@
-import React, { FC } from 'react';
+import React, { FC, memo } from 'react';
 import { Flex, Text, Grid, Button } from '@fluentui/react-northstar';
 import { useDispatch } from 'react-redux';
 
+import ConfirmDialog from '../ConfirmDialog';
 import ImageLink from '../ImageLink';
 import { deleteCamera } from '../../store/camera/cameraActions';
 
@@ -15,8 +16,8 @@ const CameraDetailInfo: FC<CameraDetailInfoProps> = ({ id, name, rtsp, modelName
   const dispatch = useDispatch();
 
   return (
-    <Flex styles={{ padding: '1rem 2rem' }} column space="between">
-      <Grid columns="2" styles={{ gap: '3rem' }}>
+    <Flex styles={{ padding: '1em 2em' }} column space="between">
+      <Grid columns="2" styles={{ gap: '3em' }}>
         <Text size="larger" weight="semibold">
           Details
         </Text>
@@ -32,11 +33,25 @@ const CameraDetailInfo: FC<CameraDetailInfoProps> = ({ id, name, rtsp, modelName
           <Text size="large" content={modelName} />
         </Flex>
       </Grid>
-      <Button
-        primary
+      <ConfirmDialog
+        trigger={
+          <Button
+            primary
+            content="Delete Camera"
+            styles={{
+              backgroundColor: '#C4314B',
+              ':hover': { backgroundColor: '#A72037' },
+              ':active': { backgroundColor: '#8E192E' },
+            }}
+          />
+        }
         content="Delete Camera"
-        styles={{ backgroundColor: '#C4314B', ':hover': { backgroundColor: '#A72037' } }}
-        onClick={(): void => {
+        primaryStyles={{
+          backgroundColor: '#C4314B',
+          ':hover': { backgroundColor: '#A72037' },
+          ':active': { backgroundColor: '#8E192E' },
+        }}
+        onConfirm={(): void => {
           dispatch(deleteCamera(id));
         }}
       />
@@ -44,4 +59,4 @@ const CameraDetailInfo: FC<CameraDetailInfoProps> = ({ id, name, rtsp, modelName
   );
 };
 
-export default CameraDetailInfo;
+export default memo(CameraDetailInfo);
